--- src/main/java/at/hannibal2/skyhanni/features/garden/fortuneguide/FFStats.kt
@@ -1,219 +1,0 @@
-package at.hannibal2.skyhanni.features.garden.fortuneguide
-
-import at.hannibal2.skyhanni.data.CropAccessoryData
-import at.hannibal2.skyhanni.data.ProfileStorageData
-import at.hannibal2.skyhanni.data.garden.CropUpgrades.getUpgradeLevel
-import at.hannibal2.skyhanni.features.garden.CropType
-import at.hannibal2.skyhanni.features.garden.FarmingFortuneDisplay
-import at.hannibal2.skyhanni.features.garden.GardenApi
-import at.hannibal2.skyhanni.features.inventory.attribute.AttributeShardsData
-import at.hannibal2.skyhanni.utils.ItemUtils.getInternalName
-import at.hannibal2.skyhanni.utils.NeuInternalName.Companion.toInternalName
-import at.hannibal2.skyhanni.utils.SimpleTimeMark
-import at.hannibal2.skyhanni.utils.SkyBlockItemModifierUtils.getFarmingForDummiesCount
-import at.hannibal2.skyhanni.utils.SkyBlockItemModifierUtils.getHeldPetItem
-import at.hannibal2.skyhanni.utils.SkyBlockItemModifierUtils.getPetLevel
-import at.hannibal2.skyhanni.utils.SkyBlockTime
-import net.minecraft.world.item.ItemStack
-import kotlin.math.floor
-
-object FFStats {
-
-    private val farmingBoots = setOf("RANCHERS_BOOTS", "FARMER_BOOTS")
-    private val GREEN_BANDANA_ITEM = "GREEN_BANDANA".toInternalName()
-    private val YELLOW_BANDANA_ITEM = "YELLOW_BANDANA".toInternalName()
-    private val MINOS_RELIC_ITEM = "MINOS_RELIC".toInternalName()
-
-    private const val SHARD_FIREFLY = "SHARD_FIREFLY"
-    private const val SHARD_LUNAR_MOTH = "SHARD_LUNAR_MOTH"
-    private const val SHARD_TERMITE = "SHARD_TERMITE"
-    private const val SHARD_GALAXY_FISH = "SHARD_GALAXY_FISH"
-
-    var cakeExpireTime
-        get() = GardenApi.storage?.fortune?.cakeExpiring ?: SimpleTimeMark.farPast()
-        set(value) {
-            GardenApi.storage?.fortune?.cakeExpiring = value
-        }
-
-    var equipmentTotalFF = mapOf<FFTypes, Double>()
-
-    var armorTotalFF = mapOf<FFTypes, Double>()
-    var usingSpeedBoots = false
-
-    var currentPetItem = ""
-
-    var baseFF = mapOf<FFTypes, Double>()
-
-    var totalBaseFF = mapOf<FFTypes, Double>()
-
-    fun loadFFData() {
-        equipmentTotalFF = FarmingItemType.equip.getFFData()
-
-        armorTotalFF = FarmingItemType.armor.getFFData()
-        usingSpeedBoots = FarmingItemType.BOOTS.getItem().getInternalName().asString() in farmingBoots
-
-        baseFF = getGenericFF()
-
-        getTotalFF()
-    }
-
-    fun getCropStats(crop: CropType, tool: ItemStack?) {
-        FortuneStats.reset()
-
-        FortuneStats.BASE.set(FFInfos.UNIVERSAL.current, FFInfos.UNIVERSAL.max)
-        FortuneStats.CROP_UPGRADE.set((crop.getUpgradeLevel()?.toDouble() ?: 0.0) * 5.0, 45.0)
-        FortuneStats.ACCESSORY.set(CropAccessoryData.cropAccessory.getFortune(crop), 30.0)
-        FortuneStats.FFD.set((tool?.getFarmingForDummiesCount() ?: 0).toDouble(), 5.0)
-        FortuneStats.TURBO.set(FarmingFortuneDisplay.getTurboCropFortune(tool, crop), 25.0)
-        FortuneStats.DEDICATION.set(FarmingFortuneDisplay.getDedicationFortune(tool, crop), 92.0)
-        FortuneStats.CULTIVATING.set(FarmingFortuneDisplay.getCultivatingFortune(tool), 20.0)
-
-        FarmingFortuneDisplay.loadFortuneLineData(tool, 0.0)
-
-        when (crop) {
-            CropType.WHEAT, CropType.CARROT, CropType.POTATO, CropType.SUGAR_CANE, CropType.NETHER_WART,
-            CropType.SUNFLOWER, CropType.MOONFLOWER, CropType.WILD_ROSE,
-            -> {
-                FortuneStats.BASE_TOOL.set(FarmingFortuneDisplay.getToolFortune(tool), 50.0)
-                FortuneStats.HARVESTING.set(FarmingFortuneDisplay.getHarvestingFortune(tool), 75.0)
-                FortuneStats.COLLECTION.set(FarmingFortuneDisplay.getCollectionFortune(tool), 48.0)
-                FortuneStats.REFORGE.set(FarmingFortuneDisplay.reforgeFortune, 20.0)
-                FortuneStats.GEMSTONE.set(FarmingFortuneDisplay.gemstoneFortune, 30.0)
-            }
-
-            CropType.PUMPKIN, CropType.MELON -> {
-                FortuneStats.SUNDER.set(FarmingFortuneDisplay.getSunderFortune(tool), 75.0)
-                FortuneStats.REFORGE.set(FarmingFortuneDisplay.reforgeFortune, 20.0)
-                FortuneStats.GEMSTONE.set(FarmingFortuneDisplay.gemstoneFortune, 30.0)
-            }
-
-            CropType.MUSHROOM -> {
-                FortuneStats.BASE_TOOL.set(FarmingFortuneDisplay.getToolFortune(tool), 30.0)
-                FortuneStats.HARVESTING.set(FarmingFortuneDisplay.getHarvestingFortune(tool), 75.0)
-                FortuneStats.REFORGE.set(FarmingFortuneDisplay.reforgeFortune, 16.0)
-                FortuneStats.GEMSTONE.set(FarmingFortuneDisplay.gemstoneFortune, 16.0)
-            }
-
-            CropType.COCOA_BEANS -> {
-                FortuneStats.BASE_TOOL.set(FarmingFortuneDisplay.getToolFortune(tool), 20.0)
-                FortuneStats.SUNDER.set(FarmingFortuneDisplay.getSunderFortune(tool), 75.0)
-                FortuneStats.REFORGE.set(FarmingFortuneDisplay.reforgeFortune, 16.0)
-                FortuneStats.GEMSTONE.set(FarmingFortuneDisplay.gemstoneFortune, 16.0)
-            }
-
-            CropType.CACTUS -> {
-                FortuneStats.HARVESTING.set(FarmingFortuneDisplay.getHarvestingFortune(tool), 75.0)
-                FortuneStats.REFORGE.set(FarmingFortuneDisplay.reforgeFortune, 16.0)
-                FortuneStats.GEMSTONE.set(FarmingFortuneDisplay.gemstoneFortune, 16.0)
-            }
-
-            else -> {}
-        }
-        CarrolynTable.getByCrop(crop)?.let {
-            val ff = if (it.get()) 12.0 else 0.0
-            FortuneStats.CARROLYN.set(ff, 12.0)
-        }
-
-        FortuneStats.PERSONAL_BEST.set(FarmingFortuneDisplay.getPersonalBest(crop), 100.0)
-        FortuneStats.CROP_TOTAL.set(FortuneStats.getTotal())
-    }
-
-    fun getEquipmentFFData(item: ItemStack?): Map<FFTypes, Double> = buildMap {
-        FarmingFortuneDisplay.loadFortuneLineData(item, 0.0)
-        this[FFTypes.BASE] = FarmingFortuneDisplay.itemBaseFortune
-        this[FFTypes.REFORGE] = FarmingFortuneDisplay.reforgeFortune
-        this[FFTypes.ENCHANT] = FarmingFortuneDisplay.greenThumbFortune
-        this[FFTypes.ABILITY] = FarmingFortuneDisplay.getAbilityFortune(item)
-        this[FFTypes.TOTAL] = this.values.sum()
-    }
-
-    fun getArmorFFData(item: ItemStack?): Map<FFTypes, Double> = buildMap {
-        FarmingFortuneDisplay.loadFortuneLineData(item, 0.0)
-        this[FFTypes.BASE] = FarmingFortuneDisplay.itemBaseFortune
-        this[FFTypes.REFORGE] = FarmingFortuneDisplay.reforgeFortune
-        this[FFTypes.GEMSTONE] = FarmingFortuneDisplay.gemstoneFortune
-        this[FFTypes.ENCHANT] = FarmingFortuneDisplay.pesterminatorFortune
-        this[FFTypes.ABILITY] = FarmingFortuneDisplay.getAbilityFortune(item)
-        this[FFTypes.TOTAL] = this.values.sum()
-    }
-
-    fun getPetFFData(item: ItemStack?): Map<FFTypes, Double> = buildMap {
-        val gardenLvl = GardenApi.getGardenLevel(overflow = false)
-        this[FFTypes.BASE] = getPetFF(item)
-        this[FFTypes.PET_ITEM] = when (item?.getHeldPetItem()) {
-            GREEN_BANDANA_ITEM -> 4.0 * gardenLvl
-            YELLOW_BANDANA_ITEM -> 30.0
-            MINOS_RELIC_ITEM -> (this[FFTypes.BASE] ?: 0.0) * .33
-            else -> 0.0
-        }
-        this[FFTypes.TOTAL] = this.values.sum()
-    }
-
-    private fun getGenericFF(): Map<FFTypes, Double> = buildMap {
-        val storage = GardenApi.storage?.fortune ?: return emptyMap()
-        this[FFTypes.FARMING_LVL] = storage.farmingLevel.toDouble() * 4
-        this[FFTypes.ATTRIBUTE_SHARDS] = getAttributeShards()
-        this[FFTypes.BESTIARY] = storage.bestiary
-        this[FFTypes.PLOTS] = storage.plotsUnlocked.toDouble() * 3
-        this[FFTypes.ANITA] = storage.anitaUpgrade.toDouble() * 4
-        this[FFTypes.COMMUNITY_SHOP] = (ProfileStorageData.playerSpecific?.gardenCommunityUpgrade ?: -1).toDouble() * 4
-        this[FFTypes.RELIC_OF_POWER] = storage.relicOfPower
-        this[FFTypes.DARK_CACAO_TRUFFLE] = storage.cacao.toDouble()
-
-        if (cakeExpireTime.isInFuture() || cakeExpireTime.isFarPast()) {
-            this[FFTypes.CAKE] = 5.0
-        } else {
-            this[FFTypes.CAKE] = 0.0
-        }
-        this[FFTypes.TOTAL] = this.values.sum()
-    }
-
-    fun getAttributeShards(): Double {
-        val solarPower = AttributeShardsData.getActiveLevel(SHARD_FIREFLY).toDouble() * 5
-        val lunarPower = AttributeShardsData.getActiveLevel(SHARD_LUNAR_MOTH).toDouble() * 5
-        val infiltration = AttributeShardsData.getActiveLevel(SHARD_TERMITE).toDouble() * 3
-        val ultimateDna = AttributeShardsData.getActiveLevel(SHARD_GALAXY_FISH).toDouble()
-        return maxOf(solarPower, lunarPower) + infiltration + ultimateDna
-    }
-
-    fun getTotalFF() {
-        currentPetItem = FarmingItemType.currentPet.getItem().getHeldPetItem().toString()
-
-        totalBaseFF = combineFFData(
-            baseFF, armorTotalFF, equipmentTotalFF, FarmingItemType.currentPet.getFFData(),
-        )
-
-        FFGuideGui.updateDisplay()
-    }
-
-    private fun List<FarmingItemType>.getFFData(): Map<FFTypes, Double> = combineFFData(this.map { it.getFFData() })
-
-    private fun combineFFData(vararg value: Map<FFTypes, Double>) = combineFFData(value.toList())
-    private fun combineFFData(value: List<Map<FFTypes, Double>>) =
-        value.map { it.toList() }.flatten().groupBy({ it.first }, { it.second })
-            .mapValues { (_, values) -> values.sum() }
-
-    private fun getPetFF(pet: ItemStack?): Double {
-        if (pet == null) return 0.0
-        val petLevel = pet.getPetLevel()
-        val strength = (GardenApi.storage?.fortune?.farmingStrength)
-        if (strength != null) {
-            val rawInternalName = pet.getInternalName()
-            return when {
-                rawInternalName.contains("ELEPHANT;4") -> 1.5 * petLevel
-                rawInternalName.contains("MOOSHROOM_COW;4") -> {
-                    (10 + petLevel).toDouble() + floor(floor(strength / (40 - petLevel * .2)) * .7)
-                }
-
-                rawInternalName.contains("MOOSHROOM") -> (10 + petLevel).toDouble()
-                rawInternalName.contains("BEE;2") -> 0.2 * petLevel
-                rawInternalName.contains("BEE;3") || rawInternalName.contains("BEE;4") -> 0.3 * petLevel
-                rawInternalName.contains("SLUG;4") -> 1.0 * petLevel
-                rawInternalName.contains("HEDGEHOG;4") -> 0.45 * petLevel * if (SkyBlockTime.isDay()) 1.0 else 3.0
-                else -> 0.0
-            }
-        }
-        return 0.0
-    }
-
-}
