--- src/main/java/at/hannibal2/skyhanni/features/garden/fortuneguide/FFTypes.kt
@@ -1,20 +1,0 @@
-package at.hannibal2.skyhanni.features.garden.fortuneguide
-
-enum class FFTypes {
-    TOTAL,
-    BASE,
-    REFORGE,
-    GEMSTONE,
-    ABILITY,
-    ENCHANT,
-    FARMING_LVL,
-    ATTRIBUTE_SHARDS,
-    BESTIARY,
-    PLOTS,
-    ANITA,
-    COMMUNITY_SHOP,
-    DARK_CACAO_TRUFFLE,
-    RELIC_OF_POWER,
-    CAKE,
-    PET_ITEM,
-}
