--- src/main/java/at/hannibal2/skyhanni/features/garden/fortuneguide/FarmingSet.kt
@@ -1,16 +1,0 @@
-package at.hannibal2.skyhanni.features.garden.fortuneguide
-
-enum class FarmingSet(
-    val helmetIncrease: Double,
-    val chestplateIncrease: Double,
-    val leggingsIncrease: Double,
-    val bootsIncrease: Double,
-) {
-
-    MELON(15.0, 20.0, 20.0, 15.0),
-    CROPIE(15.0, 20.0, 20.0, 15.0),
-    SQUASH(15.0, 20.0, 20.0, 15.0),
-    FERMENTO(15.0, 20.0, 20.0, 15.0),
-    FARMER(0.0, 0.0, 0.0, 0.0),
-    RANCHERS(0.0, 0.0, 0.0, 0.0)
-}
